Refactor Tower and reorganize about us route:
